% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igram.R
\name{igram}
\alias{igram}
\title{Return an icon}
\usage{
igram(icon, lib = "simple", size = 128, color = "default")
}
\arguments{
\item{icon}{icon name}

\item{lib}{icon library}

\item{size}{icon size}

\item{color}{icon color as hexcode, excluding #}
}
\description{
Return an icon
}
