% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igrams}
\alias{igrams}
\title{Full list of icons}
\format{
tibble with columns
\itemize{
\item lib: icon library
\item icon: icon name
\item url: icon url
}
}
\source{
\href{https://icongr.am/}{Icongram}
}
\usage{
igrams
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Full list of icons
}
\keyword{datasets}
