% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igram.R
\name{igram}
\alias{igram}
\title{Return an icon}
\usage{
igram(icon, lib = "simple", size = 128, color = "default", raw = FALSE)
}
\arguments{
\item{icon}{icon name}

\item{lib}{icon library}

\item{size}{icon size}

\item{color}{icon color as hexcode, excluding #}

\item{raw}{return raw html instead of viewing icon in rstudio viewer}
}
\description{
Return an icon
}
\seealso{
\code{\link{igrams}}
}
